test(pandas): move todate tests
